Support args from Lambda request
